// src/common/enums/analytics.enum.ts
export enum LineChartReportType {
  NUMBER_OF_CAR = 'number_of_car',
  VIEW = 'view',
  IMPRESSION = 'impression',
  CLICK = 'click',
}

export enum AnalyticsCarChartType {
  HOUR = 'hour',
  DAY = 'day',
  MONTH = 'month',
  YEAR = 'year',
}

export enum AnalyticsViewType {
  CAR = 'car',
  BANNER = 'banner',
}

export enum AnalyticsClickType {
  CAR = 'car',
  BANNER = 'banner',
  LINE = 'line',
  CHAT = 'chat',
  CALL = 'call',
}

export enum AnalyticsImpressionType {
  CAR = 'car',
  BANNER = 'banner',
}